ref(dynamic-sampling): Remove multiplexer
