+++ src/sentry/api/endpoints/project_details.py
@@ -22,8 +22,12 @@ from sentry.api.serializers.rest_framework.origin import OriginField
 from sentry.auth.superuser import is_active_superuser
 from sentry.constants import RESERVED_PROJECT_SLUGS
 from sentry.datascrubbing import validate_pii_config_update
-from sentry.dynamic_sampling.feature_multiplexer import DynamicSamplingFeatureMultiplexer
 from sentry.dynamic_sampling.rules_generator import generate_rules
+from sentry.dynamic_sampling.utils import (
+    get_supported_biases_ids,
+    get_user_biases,
+    is_on_dynamic_sampling,
+)
 from sentry.grouping.enhancer import Enhancements, InvalidEnhancerConfig
 from sentry.grouping.fingerprinting import FingerprintingRules, InvalidFingerprintingConfig
 from sentry.ingest.inbound_filters import FilterTypes
@@ -68,9 +72,7 @@ def clean_newline_inputs(value, case_insensitive=True):
 
 
 class DynamicSamplingBiasSerializer(serializers.Serializer):
-    id = serializers.ChoiceField(
-        required=True, choices=DynamicSamplingFeatureMultiplexer.get_supported_biases_ids()
-    )
+    id = serializers.ChoiceField(required=True, choices=get_supported_biases_ids())
     active = serializers.BooleanField(default=False)
 
     def validate(self, data):
@@ -371,14 +373,10 @@ class ProjectDetailsEndpoint(ProjectEndpoint):
         if "hasAlertIntegration" in expand:
             data["hasAlertIntegrationInstalled"] = has_alert_integration(project)
 
-        ds_feature_multiplexer = DynamicSamplingFeatureMultiplexer(project)
-
         # Dynamic Sampling Logic
-        if ds_feature_multiplexer.is_on_dynamic_sampling:
+        if is_on_dynamic_sampling(project):
             ds_bias_serializer = DynamicSamplingBiasSerializer(
-                data=ds_feature_multiplexer.get_user_biases(
-                    project.get_option("sentry:dynamic_sampling_biases", None)
-                ),
+                data=get_user_biases(project.get_option("sentry:dynamic_sampling_biases", None)),
                 many=True,
             )
             if not ds_bias_serializer.is_valid():
@@ -433,8 +431,7 @@ class ProjectDetailsEndpoint(ProjectEndpoint):
 
         result = serializer.validated_data
 
-        ds_flags_multiplexer = DynamicSamplingFeatureMultiplexer(project)
-        if result.get("dynamicSamplingBiases") and not ds_flags_multiplexer.is_on_dynamic_sampling:
+        if result.get("dynamicSamplingBiases") and not is_on_dynamic_sampling(project):
             return Response(
                 {"detail": ["dynamicSamplingBiases is not a valid field"]},
                 status=403,
@@ -605,9 +602,7 @@ class ProjectDetailsEndpoint(ProjectEndpoint):
             )
 
         if "dynamicSamplingBiases" in result:
-            updated_biases = ds_flags_multiplexer.get_user_biases(
-                user_set_biases=result["dynamicSamplingBiases"]
-            )
+            updated_biases = get_user_biases(user_set_biases=result["dynamicSamplingBiases"])
             if project.update_option("sentry:dynamic_sampling_biases", updated_biases):
                 changed_proj_settings["sentry:dynamic_sampling_biases"] = result[
                     "dynamicSamplingBiases"
@@ -745,7 +740,7 @@ class ProjectDetailsEndpoint(ProjectEndpoint):
         )
 
         data = serialize(project, request.user, DetailedProjectSerializer())
-        if not ds_flags_multiplexer.is_on_dynamic_sampling:
+        if not is_on_dynamic_sampling(project):
             data["dynamicSamplingBiases"] = None
         # If here because the case of when no dynamic sampling is enabled at all, you would want to kick
         # out both keys actually
