+++ src/sentry/event_manager.py
@@ -53,8 +53,8 @@ from sentry.constants import (
     DataCategory,
 )
 from sentry.culprit import generate_culprit
-from sentry.dynamic_sampling.feature_multiplexer import DynamicSamplingFeatureMultiplexer
 from sentry.dynamic_sampling.latest_release_booster import LatestReleaseBias, LatestReleaseParams
+from sentry.dynamic_sampling.utils import is_on_dynamic_sampling
 from sentry.eventstore.processing import event_processing_store
 from sentry.eventtypes import (
     CspEvent,
@@ -911,9 +911,7 @@ def _get_or_create_release_many(jobs: Sequence[Job], projects: ProjectsMapping) 
                 # Dynamic Sampling - Boosting latest release functionality
                 if (
                     options.get("dynamic-sampling:boost-latest-release")
-                    and DynamicSamplingFeatureMultiplexer(
-                        project=projects[project_id]
-                    ).is_on_dynamic_sampling
+                    and is_on_dynamic_sampling(projects[project_id])
                     and data.get("type") == "transaction"
                 ):
                     with sentry_sdk.start_span(
