+++ tests/sentry/dynamic_sampling/test_generate_rules.py
@@ -45,9 +45,7 @@ def test_generate_rules_capture_exception(get_blended_sample_rate, sentry_sdk):
     validate_sampling_configuration(config_str)
 
 
-@patch(
-    "sentry.dynamic_sampling.feature_multiplexer.DynamicSamplingFeatureMultiplexer.get_enabled_user_biases"
-)
+@patch("sentry.dynamic_sampling.utils.get_enabled_user_biases")
 @patch("sentry.dynamic_sampling.rules_generator.quotas.get_blended_sample_rate")
 def test_generate_rules_return_uniform_rules_with_rate(
     get_blended_sample_rate, get_enabled_user_biases
